Rename file and fix references
